requirements.txt
flask
twitter
python-memcached

